Fix writing specialized segments to memory
By implementing the writing of `CString`s from a `poke` instance method
we are forcing ourselves to free the memory for those strings after we
finish writing the segment, but before we can use the segment.

Instead of using `poke` we can define a custom function which can return
a `FreeAfter` type instead of `IO`. This means we write all the
components of the segment but don't clean up anything yet. The caller
can use the in-memory segments, and free them later by an explicit call
to `freeAfter`.